requirements.txt
appnope==0.1.3
asttokens==2.4.1
cffi==1.16.0
click==8.1.7
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
exif==1.6.0
h3==3.7.6
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.3
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.0.0
platformdirs==4.2.0
plum-py==0.8.7
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.2
six==1.16.0
stack-data==0.6.3
timezonefinder==6.4.0
tornado==6.4
traitlets==5.14.1
wcwidth==0.2.13
